fix: ensure timing things are not mocked (#305)
* feature/use-original-setTimeout-when-used-in-a-test

* also utilize the original clearTimeout, added tests for getSetTimeout and getClearTimeout

* added extra check to getSetTimeout and getClearTimeout to make sure that environments withouth window will work as expected

* simplified getSetTimeout and getClearTimeout

* updated tests for the Node environment, removed dependency injection for window

* updated test name

* Update src/helpers.js


* Update src/helpers.js


* updated with feedback from Kent

* do some fancy magic stuff

* no longer needed

* fix coverage
