+++ src/wait-for-dom-change.js
@@ -1,4 +1,10 @@
-import {newMutationObserver, getDocument, getSetImmediate} from './helpers'
+import {
+  newMutationObserver,
+  getDocument,
+  setImmediate,
+  setTimeout,
+  clearTimeout,
+} from './helpers'
 import {getConfig} from './config'
 
 function waitForDomChange({
@@ -12,7 +18,6 @@ function waitForDomChange({
   },
 } = {}) {
   return new Promise((resolve, reject) => {
-    const setImmediate = getSetImmediate()
     const timer = setTimeout(onTimeout, timeout)
     const observer = newMutationObserver(onMutation)
     observer.observe(container, mutationObserverOptions)
