+++ src/wait-for-element.js
@@ -1,4 +1,10 @@
-import {newMutationObserver, getDocument, getSetImmediate} from './helpers'
+import {
+  newMutationObserver,
+  getDocument,
+  setImmediate,
+  setTimeout,
+  clearTimeout,
+} from './helpers'
 import {getConfig} from './config'
 
 function waitForElement(
@@ -23,10 +29,10 @@ function waitForElement(
     }
     let lastError
     const timer = setTimeout(onTimeout, timeout)
+
     const observer = newMutationObserver(onMutation)
     observer.observe(container, mutationObserverOptions)
     function onDone(error, result) {
-      const setImmediate = getSetImmediate()
       clearTimeout(timer)
       setImmediate(() => observer.disconnect())
       if (error) {
