+++ tests/jest.config.node.js
@@ -4,7 +4,6 @@ const baseConfig = require('kcd-scripts/jest')
 module.exports = {
   ...baseConfig,
   rootDir: path.join(__dirname, '..'),
-  setupFilesAfterEnv: undefined,
   displayName: 'node',
   testEnvironment: 'jest-environment-node',
   testMatch: ['**/__node_tests__/**.js'],
